+++ docs/changelog.rst
@@ -6,6 +6,11 @@ Changelog
 `Unreleased`_
 -------------
 
+Fixed
+~~~~~
+
+- Fix Bcc ignoring in Django backend. `#135`_
+
 `0.10.0`_ - 2017-03-30
 ----------------------
 
@@ -267,6 +272,7 @@ Fixed
 .. _0.2.0: https://github.com/Stranger6667/postmarker/compare/0.1.1...0.2.0
 .. _0.1.1: https://github.com/Stranger6667/postmarker/compare/0.1.0...0.1.1
 
+.. _#135: https://github.com/Stranger6667/postmarker/issues/135
 .. _#130: https://github.com/Stranger6667/postmarker/issues/130
 .. _#128: https://github.com/Stranger6667/postmarker/issues/128
 .. _#125: https://github.com/Stranger6667/postmarker/issues/125